pyproject.toml
[tool.poetry]
name = "dbt_test_coverage"
version = "0.1.0"
description = "Comprehensive test coverage analysis for dbt projects"
authors = ["Chinmayee Udupa <chinmayeeudupa023@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/chinmayee-udupa/dbt-test-coverage"
repository = "https://github.com/chinmayee-udupa/dbt-test-coverage"
documentation = "https://github.com/chinmayee-udupa/dbt-test-coverage"
packages = [{ include = "dbt_test_coverage", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
dbt-core = "^1.8.0"
dbt-postgres = "^1.8.0"
rich = "^10.0"
click = "^8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7"
pytest-docker = "^2"
black = "==24.2.0"
isort = "==5.13.2"
pre-commit = "^3.6.2"
tox = "^4"
typing-extensions = "*"
importlib-metadata = "*"

[tool.poetry.scripts]
dbt-test-coverage = "dbt_test_coverage.cli:main"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
